requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.22
pydantic==1.10.8
python-dotenv==1.0.0
requests==2.31.0
python-jose==3.3.0
